Refactor: move DropDown types to a separate file and update imports
