--- src/components/cms/ui/DropDown/DropDownOption.ts
@@ -1,7 +1,0 @@
-export type DropDownOption = {
-  label: string
-  value: string
-  icon?: React.ReactNode
-  onClick?: () => void
-  isDivider?: boolean
-}
