+++ src/components/cms/ui/DropDown/index.ts
@@ -1,2 +1,7 @@
-export { default } from './DropDown'
-export type { DropDownOption } from './DropDownOption'
+export { default as DropDown } from './DropDown'
+export type {
+  DropDownProps,
+  DropDownButtonProps,
+  DropDownListProps,
+  DropDownOption,
+} from './DropDown.types'
